requirements.txt
click==7.1.2
fastapi==0.61.2
h11==0.11.0
Pillow==8.0.1
pydantic==1.7.2
python-multipart==0.0.5
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.12.2
